move 'Hint' component to 'src/components' folder
