+++ src/components/Hint/Hint.scss
