+++ src/modules/forms/plugins/Cors/CorsPlugin.js
@@ -10,7 +10,7 @@ import getValues from '../../../../helpers/getValues';
 import Row from '../../../Layout/Row/Row';
 import Label from '../../../labels/Label';
 import Input from '../../../inputs/Input';
-import Hint from '../../../labels/Hint/Hint';
+import Hint from '../../../../components/Hint/Hint';
 import ControlBar from '../ControlBar/ControlBar';
 import MultiSelect from '../../../selects/MultiSelect/MultiSelect';
 import TagSelect from '../../../selects/TagSelect/TagSelect';
