+++ src/modules/forms/plugins/RateLimit/RateLimitPlugin.js
@@ -8,7 +8,7 @@ import block from '../../../../helpers/bem-cn';
 import Row from '../../../Layout/Row/Row';
 import Label from '../../../labels/Label';
 import Input from '../../../inputs/Input';
-import Hint from '../../../labels/Hint/Hint';
+import Hint from '../../../../components/Hint/Hint';
 import ControlBar from '../ControlBar/ControlBar';
 import SimpleSelect from '../../../selects/SimpleSelect/SimpleSelect';
