+++ src/modules/forms/plugins/RequestTransformer/HeadersSection/HeadersSection.js
@@ -7,7 +7,7 @@ import block from '../../../../../helpers/bem-cn';
 import Row from '../../../../Layout/Row/Row';
 import Label from '../../../../labels/Label';
 import Input from '../../../../inputs/Input';
-import Hint from '../../../../labels/Hint/Hint';
+import Hint from '../../../../../components/Hint/Hint';
 import Control from '../../../../../components/Control/Control';
 
 const row = block('j-row');
